streamlit-dashboard/requirements.txt
streamlit
paho-mqtt
pandas
numpy
requests
lime
plotly
matplotlib
scikit-learn
onnxruntime
streamlit-autorefresh
